add irc channel and telegram group for gsoc-cn
One thing needs to be been clarify in here is there is a gsoc-cn channel on freenode. And it registered on Apr 25 2009:

>Information on #gsoc-cn:
>3:16 PM Founder    : freenode-staff
>3:16 PM Registered : Apr 25 08:00:19 2009 (8y 20w 1d ago)
>3:16 PM Last used  : Oct 29 14:26:30 2010 (6y 45w 4d ago)
>3:16 PM Mode lock  : +ntc-slk

I have no way to find who create that channel and ask him to give me the admin permission. So, I decide to name it as gsoccn rather than gsoc-cn.